reduce kernel32_detour example
