+++ Cargo.toml
@@ -26,7 +26,6 @@ slice-pool = "0.4.1"
 [dev-dependencies]
 matches = "0.1.10"
 ctor = "0.2.0"
-minidl = "0.1.5"
 once_cell = "1.17.1"
 
 [features]
@@ -38,7 +37,11 @@ name = "messageboxw_detour"
 crate-type = ["cdylib"]
 
 [[example]]
-name = "cat-detour"
+name = "cat_detour"
+crate-type = ["cdylib"]
+
+[[example]]
+name = "kernel32_detour"
 crate-type = ["cdylib"]
 
 [target."cfg(any(target_arch = \"x86\", target_arch = \"x86_64\"))".dependencies]
