+++ examples/cat_detour.rs
